// types.ts
export interface Band {
  id: string;
  name: string;
  energy: number; // kWh
  amount: number; // ARS
}
// ...
export interface CoopResidencialData {
  energiaGenerada: number;
  energiaInyectada: number;
  energiaEntregada: number;
  cargosFijos: number[];
  reconocimientoAmbiental: number;
  bands: Band[];
  subtotalEnergia: number;
  totalPagar: number;
  subtotalImpuestos: number;
}

export type CoopComercialData = CoopResidencialData;
export type CoopIndustrialData = CoopResidencialData;